pyproject.toml
[tool.poetry]
name = "lazaza-assessment"
version = "0.1.0"
description = ""
authors = ["Brent Quackenbush <brentprogramming@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pyyaml = "^6.0.1"
aiohttp = "^3.9.5"
jsonschema = "^4.22.0"
pydantic = "^2.7.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
